+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/messages/list/adapter/viewholder/decorator/internal/FailedIndicatorDecorator.kt
@@ -21,6 +21,7 @@ import androidx.core.view.isVisible
 import io.getstream.chat.android.ui.feature.messages.list.MessageListItemStyle
 import io.getstream.chat.android.ui.feature.messages.list.adapter.MessageListItem
 import io.getstream.chat.android.ui.feature.messages.list.adapter.viewholder.decorator.BaseDecorator
+import io.getstream.chat.android.ui.feature.messages.list.adapter.viewholder.decorator.Decorator
 import io.getstream.chat.android.ui.feature.messages.list.adapter.viewholder.impl.CustomAttachmentsViewHolder
 import io.getstream.chat.android.ui.feature.messages.list.adapter.viewholder.impl.FileAttachmentsViewHolder
 import io.getstream.chat.android.ui.feature.messages.list.adapter.viewholder.impl.GiphyAttachmentViewHolder
@@ -41,6 +42,11 @@ internal class FailedIndicatorDecorator(
     private val listViewStyle: MessageListItemStyle,
     private val isCurrentUserBanned: () -> Boolean,
 ) : BaseDecorator() {
+
+    /**
+     * The type of the decorator. In this case [Decorator.Type.BuiltIn.FAILED_INDICATOR].
+     */
+    override val type: Decorator.Type = Decorator.Type.BuiltIn.FAILED_INDICATOR
 
     /**
      * Decorates the visibility of the "failed" section of the message containing