+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/messages/list/adapter/viewholder/decorator/internal/ReactionsDecorator.kt
@@ -26,6 +26,7 @@ import androidx.core.view.updateLayoutParams
 import io.getstream.chat.android.ui.feature.messages.list.MessageListItemStyle
 import io.getstream.chat.android.ui.feature.messages.list.adapter.MessageListItem
 import io.getstream.chat.android.ui.feature.messages.list.adapter.viewholder.decorator.BaseDecorator
+import io.getstream.chat.android.ui.feature.messages.list.adapter.viewholder.decorator.Decorator
 import io.getstream.chat.android.ui.feature.messages.list.adapter.viewholder.impl.CustomAttachmentsViewHolder
 import io.getstream.chat.android.ui.feature.messages.list.adapter.viewholder.impl.FileAttachmentsViewHolder
 import io.getstream.chat.android.ui.feature.messages.list.adapter.viewholder.impl.GiphyAttachmentViewHolder
@@ -41,6 +42,11 @@ import io.getstream.chat.android.ui.utils.extensions.hasSingleReaction
 import io.getstream.chat.android.ui.utils.extensions.updateConstraints
 
 internal class ReactionsDecorator(private val style: MessageListItemStyle) : BaseDecorator() {
+
+    /**
+     * The type of the decorator. In this case [Decorator.Type.BuiltIn.REACTIONS].
+     */
+    override val type: Decorator.Type = Decorator.Type.BuiltIn.REACTIONS
 
     /**
      * Decorates the reactions section of the message containing custom attachments.